backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
drf-spectacular
Pillow
pytest
pytest-django
model-bakery
